# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Archery Manager mobile API
  version: 1.0.0
  summary: Stateless JSON API consumed by the Archery Manager mobile app.
  description: |
    This is the contract between the Symfony application and the React Native app. The app generates its
    typed client from this file; `tests/Integration/Api/OpenApiSpecTest.php` fails when a route and this
    document drift apart.

    ## Conventions

    - **Versioned and stateless.** Everything lives under `/api/v1`. There is no session and no cookie.
    - **Technical text is English**: error codes, error messages, field names. **Display labels are French**
      (`label` of an enum value, club and group names, event titles...): they are the club's own content.
    - **Bearer tokens.** `POST /auth/login` returns an opaque access token (15 minutes) and a refresh token
      (30 days). Send `Authorization: Bearer <access_token>`. A refresh token works **once**: refreshing
      returns a new pair and the previous refresh token becomes invalid.
    - **Request context.** One account can have several licensees (family members) and the data depends on a
      season. Both are chosen per request with the `X-Licensee` and `X-Season` headers, never stored on the
      server. Without them the first licensee and the current season are used.
    - **Errors** always have the same shape, see `Error`, and carry a stable machine-readable `error` code.
...

    ## Access rules and their known limit

...

...

...
  license:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /club:
    get:
      tags: [Club]
      operationId: getClub
      summary: The club of the selected licensee for the selected season
      description: Needs a license for the selected season (`403 forbidden` otherwise).
      parameters:
        - $ref: '#/components/parameters/XLicensee'
        - $ref: '#/components/parameters/XSeason'
      responses:
        '200':
          description: The club, its groups and member counts.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClubOverview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /club/members:
    get:
      tags: [Club]
      operationId: listClubMembers
      summary: The member directory ("trombinoscope")
      description: |
        The members of the club that hold a license for the selected season, sorted by displayed name
        (case and accents ignored). Needs a license for the selected season.

        **Privacy**: admins and coaches see full names; everybody else sees "Firstname L.". The search `q`
        only looks at the displayed name, so it cannot be used to find out a hidden last name.
      parameters:
        - $ref: '#/components/parameters/XLicensee'
        - $ref: '#/components/parameters/XSeason'
        - name: group
          in: query
          description: Only the members of this group (the id of a group of the club), or `none` for the members without a group. Anything else is a `400`.
          schema:
            type: string
            pattern: '^(none|[1-9][0-9]*)$'
        - name: q
          in: query
          description: Search in the displayed name, ignoring case and accents.
          schema:
            type: string
        - name: page
          in: query
          description: Starts at 1. Larger than 100000 is a `400`.
          schema:
            type: integer
            minimum: 1
            maximum: 100000
            default: 1
        - name: per_page
          in: query
          description: Larger values are not rejected, they are capped to 100.
          schema:
            type: integer
            minimum: 1
            default: 30
      responses:
        '200':
          description: One page of members.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberPage'
        '400':
          description: |
            A parameter is invalid: `page` is not an integer between 1 and 100000, `per_page` is not a
            positive integer, `group` is not a group of this club, or one of the context headers is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The opaque `access_token` returned by `/auth/login` or `/auth/refresh`.
  parameters:
    XLicensee:
      name: X-Licensee
      in: header
      description: |
        FFTA member code of the licensee the request acts as. Must be a licensee of the authenticated account
        (`403` otherwise). Defaults to the account's first licensee.
      schema:
        type: string
    XSeason:
      name: X-Season
      in: header
      description: The season to use, as the year it ends (the 2026/2027 season is `2027`). Defaults to the current season.
      schema:
        type: integer
        minimum: 2000
        maximum: 2100
...
  headers:
...
    PrivateNoStore:
      description: The answer depends on the authenticated user and is never stored.
      schema:
        type: string
        examples: ['no-store, private']
...
  responses:
...
    BadRequest:
      description: |
        The request is invalid: malformed JSON or body (`bad_request`), an invalid `X-Season` header, or an
        invalid parameter.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: |
        No token (`authentication_required`), or an unknown, expired or revoked access token (`invalid_token`),
        or a locked account (`account_locked`). The response carries a `WWW-Authenticate: Bearer` challenge.
        On `invalid_token`, refresh the tokens once and retry.
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: |
        Authenticated but not allowed (`forbidden`): no license for the selected season, a licensee that
        is not yours in `X-Licensee`, or no right on the requested licensee.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [error, message]
      additionalProperties: false
      properties:
        error:
          type: string
          description: Stable machine-readable code. Branch on this, never on `message`.
          enum:
            - authentication_required
            - invalid_token
            - invalid_credentials
            - account_locked
            - too_many_attempts
            - invalid_request
            - invalid_refresh_token
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - method_not_allowed
            - not_acceptable
            - conflict
            - payload_too_large
            - unsupported_media_type
            - unprocessable_entity
            - too_many_requests
            - service_unavailable
            - http_error
            - server_error
        message:
          type: string
          description: English, for logs and developers. Not meant to be shown to users.
...
    EnumValue:
      type: object
      description: A value of the club's domain vocabulary.
      required: [code, label]
      properties:
        code:
          type: string
          description: Stable code, e.g. `CL`.
        label:
          type: string
          description: French display label, e.g. "Classique".
    Reference:
      type: object
      required: [id, name]
      properties:
        id:
          type: [integer, 'null']
        name:
          type: [string, 'null']
...
    Club:
      type: object
      required: [id, name, city, contact_email, ffta_code, primary_color, logo_url]
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: [string, 'null']
        contact_email:
          type: [string, 'null']
        ffta_code:
          type: [string, 'null']
        primary_color:
          type: [string, 'null']
          description: Hexadecimal color, e.g. `#0a7d33`.
        logo_url:
          type: [string, 'null']
          description: Path or URL of the club logo (public, no token needed), `null` when the club has none.
    Group:
      type: object
      required: [id, name, description, member_count]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: [string, 'null']
        member_count:
          type: integer
          description: Members of the group among the club's members of the selected season.
    ClubOverview:
      type: object
      required: [club, season, member_count, groups, without_group_count]
      properties:
        club:
          $ref: '#/components/schemas/Club'
        season:
          type: integer
        member_count:
          type: integer
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
        without_group_count:
          type: integer
    MemberSummary:
      type: object
      description: A member as shown on a directory card.
      required: [id, display_name, groups, activities, picture_url]
      properties:
        id:
          type: integer
        display_name:
          type: string
          description: Full name for admins and coaches, "Firstname L." for everybody else.
        groups:
          type: array
          description: The groups of the club the member is in (groups of a previous club are not listed).
          items:
            $ref: '#/components/schemas/Reference'
        activities:
          type: array
          description: The activities of the license of the selected season.
          items:
            $ref: '#/components/schemas/EnumValue'
        picture_url:
          type: [string, 'null']
          description: Path (from the host root) of the picture endpoint, `null` when there is no picture. Needs the bearer token.
    MemberPage:
      type: object
      required: [data, meta, filters]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MemberSummary'
        meta:
          type: object
          required: [page, per_page, total, total_in_club]
          properties:
            page:
              type: integer
            per_page:
              type: integer
            total:
              type: integer
              description: Members matching the filters, over all pages.
            total_in_club:
              type: integer
              description: Members of the club, whatever the filters.
        filters:
          type: object
          description: What the filter chips of the directory need.
          required: [groups, without_group_count]
          properties:
            groups:
              type: array
              items:
                $ref: '#/components/schemas/Group'
            without_group_count:
              type: integer
...